+++ packages/discovery/src/flatten/ContractFlattener.ts
@@ -4,29 +4,11 @@ import { parse } from '@solidity-parser/parser'
 // Either we ignore this error or we fork the parser and expose the types.
 // eslint-disable-next-line import/no-unresolved
 import {
-  BaseASTNode,
   ContractDefinition,
-  CustomErrorDefinition,
-  EventDefinition,
-  Expression,
-  Block,
-  EmitStatement,
-  ExpressionStatement,
-  TypeDefinition,
-  RevertStatement,
-  IndexRangeAccess,
-  TypeName,
-  IfStatement,
-  Identifier,
-  VariableDeclarationStatement,
-  ReturnStatement,
-  VariableDeclaration,
-  ModifierDefinition,
-  StateVariableDeclaration,
-  StructDefinition,
   UsingForDeclaration,
 } from '@solidity-parser/parser/dist/src/ast-types'
 import * as path from 'path'
+import { getUniqueIdentifiers } from './astWalk'
 
 type ParseResult = ReturnType<typeof parse>
 
@@ -36,6 +18,7 @@ interface ByteRange {
 }
 
 interface ContractDecl {
+  ast: ContractDefinition,
   inheritsFrom: string[]
   librariesUsed: string[]
   name: string
@@ -115,9 +98,10 @@ export class ContractFlattener {
         assert(c.type === 'ContractDefinition' && c.range !== undefined)
 
         return {
+          ast: c,
           name: c.name,
           inheritsFrom: c.baseContracts.map((bc) => bc.baseName.namePath),
-          librariesUsed: this.resolveLibrariesUsed(file, c),
+          librariesUsed: [],
           byteRange: {
             start: c.range[0],
             end: c.range[1],
@@ -130,6 +114,13 @@ export class ContractFlattener {
     for (const file of this.files) {
       const visitedPaths: string[] = [file.path]
       file.importDirectives = this.resolveFileImports(file, visitedPaths)
+    }
+
+    // Pass 3: Resolve all libraries used
+    for (const file of this.files) {
+        for(const contract of file.contractDeclarations) {
+            contract.librariesUsed = this.resolveLibrariesUsed(file, contract.ast)
+        }
     }
   }
 
@@ -145,10 +136,11 @@ export class ContractFlattener {
       })
 
     let path: string[] = []
-    let idents = c.subNodes.flatMap((k) =>
-      searchForLibraries(k, path, file.content),
-    )
-    console.log(new Set(idents))
+    const identifiers = new Set(c.subNodes.flatMap((k) =>
+      getUniqueIdentifiers(k, path, file.content),
+    ))
+
+    console.log(identifiers)
 
     return fromUsingDirectives
   }
@@ -418,294 +410,10 @@ export class ContractFlattener {
   }
 }
 
-// function replaceAll(str: string, search: string, replacement: string): string {
-//   return str.split(search).join(replacement)
-// }
-
 function pathsMatch(path1: string, path2: string): boolean {
-  // This is better but for the demo we will use the below
   return path1 === path2
-
-  // return (
-  //   path1.endsWith(replaceAll(path.normalize(path2), '../', '')) &&
-  //   path.basename(path1) === path.basename(path2)
-  // )
 }
 
 function pushSource(acc: string, source: string, byteRange: ByteRange): string {
   return acc + source.slice(byteRange.start, byteRange.end + 1) + '\n\n'
-}
-
-function searchForLibraries(
-  node: BaseASTNode | null,
-  path: string[],
-  content: string,
-): string[] {
-  if (node === null) {
-    return []
-  }
-
-  assert(node.range !== undefined)
-  path.push(
-    '\n-----------------------------------------------------\n' +
-      content.slice(node.range[0], node.range[1] + 1) +
-      '\n-----------------------------------------------------\n',
-  )
-
-  switch (node.type) {
-    case 'Identifier': {
-      return [(node as Identifier).name]
-    }
-    case 'VariableDeclaration': {
-      // TODO(radomski): Decode the type, see if it contains a dot, if
-      // it does, check if it's a library
-      const decl = node as VariableDeclaration
-      const ident = decl.identifier !== null ? [decl.identifier.name] : []
-      const expr = parseExpression(decl.expression, path, content)
-      const typeName = parseTypeName(decl.typeName)
-      return expr.concat(ident).concat(typeName)
-    }
-    case 'Block': {
-      const block = node as Block
-      return block.statements.flatMap((statement) =>
-        searchForLibraries(statement, path, content),
-      )
-    }
-    case 'InlineAssemblyStatement': {
-      return []
-    }
-    case 'RevertStatement': {
-      const revertStatement = node as RevertStatement
-      return parseExpression(revertStatement.revertCall, path, content)
-    }
-    case 'IfStatement': {
-      const ifStatement = node as IfStatement
-      const condition = parseExpression(ifStatement.condition, path, content)
-      const trueBody = searchForLibraries(ifStatement.trueBody, path, content)
-      const falseBody = searchForLibraries(ifStatement.falseBody, path, content)
-
-      return condition.concat(trueBody).concat(falseBody)
-    }
-    case 'ExpressionStatement': {
-      const expressionStatement = node as ExpressionStatement
-      return parseExpression(expressionStatement.expression, path, content)
-    }
-    case 'CustomErrorDefinition': {
-      return (node as CustomErrorDefinition).parameters.flatMap((p) =>
-        searchForLibraries(p, path, content),
-      )
-    }
-    case 'EventDefinition': {
-      return (node as EventDefinition).parameters.flatMap((p) =>
-        searchForLibraries(p, path, content),
-      )
-    }
-    case 'FunctionDefinition': {
-      return (node as EventDefinition).parameters.flatMap((p) =>
-        searchForLibraries(p, path, content),
-      )
-    }
-    case 'ModifierDefinition': {
-      const mod = node as ModifierDefinition
-      const params = mod.parameters ?? []
-
-      const paramTypes = params.flatMap((p) =>
-        searchForLibraries(p, path, content),
-      )
-      const librariesFromBlock = searchForLibraries(mod.body, path, content)
-
-      return paramTypes.concat(librariesFromBlock)
-    }
-    case 'VariableDeclarationStatement': {
-      const declaration = node as VariableDeclarationStatement
-
-      const variables = declaration.variables.flatMap((v) =>
-        searchForLibraries(v, path, content),
-      )
-      const initialValue = parseExpression(
-        declaration.initialValue,
-        path,
-        content,
-      )
-
-      return variables.concat(initialValue)
-    }
-    case 'StateVariableDeclaration': {
-      const decl = node as StateVariableDeclaration
-
-      const varTypes = decl.variables.flatMap((v) =>
-        searchForLibraries(v, path, content),
-      )
-      const expr = parseExpression(decl.initialValue, path, content)
-
-      return expr.concat(varTypes)
-    }
-    case 'StructDefinition': {
-      return (node as StructDefinition).members.flatMap((m) =>
-        searchForLibraries(m, path, content),
-      )
-    }
-    case 'ReturnStatement': {
-      const returnStatement = node as ReturnStatement
-      return parseExpression(returnStatement.expression, path, content)
-    }
-    case 'EmitStatement': {
-      const emitStatement = node as EmitStatement
-      return parseExpression(emitStatement.eventCall, path, content)
-    }
-    case 'BinaryOperation': {
-      return parseExpression(node as Expression, path, content)
-    }
-    case 'EnumDefinition': {
-      return []
-    }
-    case 'TypeDefinition': {
-      const typeDefinition = node as TypeDefinition
-      return parseTypeName(typeDefinition.definition)
-    }
-    case 'UsingForDeclaration': {
-      // NOTE(radomski): We might actually want to use this in the future
-      return []
-    }
-    case 'IndexAccess':
-    case 'IndexRangeAccess':
-    case 'TupleExpression':
-    case 'BinaryOperation':
-    case 'Conditional':
-    case 'MemberAccess':
-    case 'FunctionCall':
-    case 'UnaryOperation':
-    case 'NewExpression':
-    case 'NameValueExpression': {
-      return parseExpression(node as Expression, path, content)
-    }
-    case 'ElementaryTypeName':
-    case 'UserDefinedTypeName':
-    case 'Mapping':
-    case 'ArrayTypeName':
-    case 'FunctionTypeName': {
-      return parseTypeName(node as TypeName)
-    }
-    default: {
-      throw new Error(
-        `TopLevelFunc: Unknown node type: [${node.type}] [${path.join(
-          ' -> \n',
-        )}]}]`,
-      )
-    }
-  }
-}
-
-function parseExpression(
-  expr: Expression | null,
-  path: string[],
-  content: string,
-): string[] {
-  if (!expr || !expr.type) {
-    return []
-  }
-
-  assert(expr.range !== undefined)
-  path.push(
-    '\n-----------------------------------------------------\n' +
-      content.slice(expr.range[0], expr.range[1] + 1) +
-      '\n-----------------------------------------------------\n',
-  )
-
-  switch (expr.type) {
-    case 'BinaryOperation': {
-      return parseExpression(expr.left, path, content).concat(
-        parseExpression(expr.right, path, content),
-      )
-    }
-    case 'FunctionCall': {
-      return parseExpression(expr.expression, path, content)
-        .concat(
-          expr.arguments.flatMap((k) => parseExpression(k, path, content)),
-        )
-        .concat(expr.identifiers.map((i) => i.name))
-    }
-    case 'IndexAccess': {
-      return parseExpression(expr.base, path, content).concat(
-        parseExpression(expr.index, path, content),
-      )
-    }
-    case 'TupleExpression': {
-      return expr.components.flatMap((component) =>
-        searchForLibraries(component, path, content),
-      )
-    }
-    case 'MemberAccess': {
-      return parseExpression(expr.expression, path, content)
-    }
-    case 'Conditional': {
-      return parseExpression(expr.condition, path, content)
-        .concat(parseExpression(expr.trueExpression, path, content))
-        .concat(parseExpression(expr.falseExpression, path, content))
-    }
-    case 'Identifier': {
-      return [expr.name]
-    }
-    case 'NewExpression': {
-      return parseTypeName(expr.typeName)
-    }
-    case 'UnaryOperation': {
-      return parseExpression(expr.subExpression, path, content)
-    }
-    case 'IndexRangeAccess': {
-      const base = parseExpression(expr.base, path, content)
-      const indexStart = parseExpression(expr.indexStart ?? null, path, content)
-      const indexEnd = parseExpression(expr.indexEnd ?? null, path, content)
-
-      return base.concat(indexStart).concat(indexEnd)
-    }
-    case 'NumberLiteral': {
-      return []
-    }
-    case 'BooleanLiteral': {
-      return []
-    }
-    case 'HexLiteral': {
-      return []
-    }
-    case 'StringLiteral': {
-      return []
-    }
-    case 'ElementaryTypeName': {
-      return parseTypeName(expr)
-    }
-    default: {
-      throw new Error(
-        `parseExpression: Unknown expression type: [${expr.type}] [${path.join(
-          ' -> \n',
-        )}]}]`,
-      )
-    }
-  }
-}
-
-function parseTypeName(type: TypeName | null): string[] {
-  if (!type || !type.type) {
-    return []
-  }
-
-  switch (type.type) {
-    case 'ElementaryTypeName': {
-      return [type.name]
-    }
-    case 'UserDefinedTypeName': {
-      // NOTE(radomski): Parse the dot here?
-      return [type.namePath]
-    }
-    case 'Mapping': {
-      return parseTypeName(type.keyType).concat(parseTypeName(type.valueType))
-    }
-    case 'ArrayTypeName': {
-      return parseTypeName(type.baseTypeName)
-    }
-    case 'FunctionTypeName': {
-      // NOTE(radomski): I think this is useless
-      return []
-    }
-  }
 }
